requirements.txt
celery==5.3.6
fastapi==0.109.2
uvicorn==0.27.0.post1
redis==5.0.1
python-dotenv==1.0.1
send-mail==0.0.2
flower==2.0.1